move multi-item scoring into a separate file
